+++ packages/api/db/migrations/002-import-metrics-cargo-schema.sql
@@ -1,0 +1,5 @@
+-- Import dag cargo schema
+IMPORT FOREIGN SCHEMA cargo
+    LIMIT TO (metrics, metrics_log)
+    FROM SERVER dag_cargo_server
+    INTO cargo;