+++ packages/website/.storybook/preview.js
@@ -1,0 +1,6 @@
+// .storybook/preview.js
+import '../styles/global.css'
+
+export const parameters = {
+  actions: { argTypesRegex: '^on[A-Z].*' },
+}