+++ packages/api/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [2.22.0](https://github.com/nftstorage/nft.storage/compare/api-v2.21.0...api-v2.22.0) (2022-04-28)
+
+
+### Features
+
+* add cargo metric statements and dummy data inserts for testing … ([#1835](https://github.com/nftstorage/nft.storage/issues/1835)) ([72480d6](https://github.com/nftstorage/nft.storage/commit/72480d6a9719308bf4a8a81c34764a51f5f58cfa))
+* cargo schema migration ([#1856](https://github.com/nftstorage/nft.storage/issues/1856)) ([fe2ce9c](https://github.com/nftstorage/nft.storage/commit/fe2ce9c3b084687104ce51d1cd972b0b957c9cea))
+
+
+### Bug Fixes
+
+* single() on multiple rows ([#1861](https://github.com/nftstorage/nft.storage/issues/1861)) ([459c4c1](https://github.com/nftstorage/nft.storage/commit/459c4c1c7df47a9e9a181cda3acefaa826a44ae3))
+* update db-types and add documentation ([#1619](https://github.com/nftstorage/nft.storage/issues/1619)) ([a3b28e2](https://github.com/nftstorage/nft.storage/commit/a3b28e2df4d09fd914a58e14fe362f4ec991279a))
+
+## [2.21.0](https://github.com/nftstorage/nft.storage/compare/api-v2.20.0...api-v2.21.0) (2022-04-21)
+
+
+### Features
+
+* add HasSuperHotAccess user tag ([#1838](https://github.com/nftstorage/nft.storage/issues/1838)) ([019a505](https://github.com/nftstorage/nft.storage/commit/019a505e8f4bb93a24b8c480646779f5e4b66326))
+* add permissions endpoint ([#1753](https://github.com/nftstorage/nft.storage/issues/1753)) ([2f5b6fb](https://github.com/nftstorage/nft.storage/commit/2f5b6fb2660231e4591f37e29389c41ac7045605))
+
+
+### Bug Fixes
+
+* pin magic admin ([#1841](https://github.com/nftstorage/nft.storage/issues/1841)) ([1bec0b0](https://github.com/nftstorage/nft.storage/commit/1bec0b03dcf13f700ceeaf2a0e3cce8f291f0072))
+* remove cluster API URL from /version endpoint ([#1843](https://github.com/nftstorage/nft.storage/issues/1843)) ([cece7ce](https://github.com/nftstorage/nft.storage/commit/cece7ce7773b9c37a18d8c0f33b0ec307474e294))
+
 ## [2.20.0](https://github.com/nftstorage/nft.storage/compare/api-v2.19.0...api-v2.20.0) (2022-04-12)